+++ src/types/user.type.ts
@@ -4,3 +4,8 @@ export type userData = {
   childName?: string; // 선택적 필드
   childGrade?: number; // 선택적 필드
 };
+
+export type UpdateChildRequest = {
+  name: string;
+  grade: number;
+};
