+++ src/utils/constants/grades.ts
@@ -1,11 +1,11 @@
 export const GRADE_OPTIONS = [
-  { label: '초등 1학년', value: '1' },
-  { label: '초등 2학년', value: '2' },
-  { label: '초등 3학년', value: '3' },
-  { label: '초등 4학년', value: '4' },
-  { label: '초등 5학년', value: '5' },
-  { label: '초등 6학년', value: '6' },
-  { label: '중등 1학년', value: '7' },
-  { label: '중등 2학년', value: '8' },
-  { label: '중등 3학년', value: '9' },
+  { label: '초등 1학년', value: 1 },
+  { label: '초등 2학년', value: 2 },
+  { label: '초등 3학년', value: 3 },
+  { label: '초등 4학년', value: 4 },
+  { label: '초등 5학년', value: 5 },
+  { label: '초등 6학년', value: 6 },
+  { label: '중등 1학년', value: 7 },
+  { label: '중등 2학년', value: 8 },
+  { label: '중등 3학년', value: 9 },
 ];
